+++ internal/component/prometheus/operator/common/crdmanager.go
@@ -146,7 +146,10 @@ func (c *crdManager) Run(ctx context.Context) error {
 
 	// Start prometheus scrape manager.
 	alloyAppendable := prometheus.NewFanout(c.args.ForwardTo, c.opts.ID, c.opts.Registerer, c.ls)
-	opts := &scrape.Options{}
+
+	opts := &scrape.Options{
+		EnableNativeHistogramsIngestion: c.args.Scrape.ScrapeNativeHistograms,
+	}
 	c.scrapeManager, err = scrape.NewManager(opts, slog.New(logging.NewSlogGoKitHandler(c.logger)), nil, alloyAppendable, unregisterer)
 	if err != nil {
 		return fmt.Errorf("creating scrape manager: %w", err)